Standardize Rust test validation
Run formatting, all-target checks, Clippy, and workspace tests through the
shared Rust gate used by CI and release preflight.

Document the placement convention, extract oversized unit test modules, and
add coverage for debugger, project, and remote behavior.
