+++ CONTRIBUTING.md
@@ -35,3 +35,14 @@ bun dev
 - Use descriptive commit messages (i.e., "Add autocompletion")
 - One logical change per commit
 - Update documentation if needed
+
+## Rust Tests
+
+- Keep small unit tests in a `#[cfg(test)] mod tests` block beside the code they exercise.
+- Move a large unit-test suite to `src/<module>/tests.rs` and declare it with
+  `#[cfg(test)] mod tests;` from the owning module.
+- Use `crates/<crate>/tests/` for integration tests that exercise the crate through its public API.
+- Add focused regression tests for behavior changes and bug fixes instead of targeting a coverage
+  percentage.
+- Run one crate or test while iterating with `cargo test -p <crate> <test-name>`, then run
+  `bun check:rust` before submitting Rust changes.
