+++ scripts/check/rust.sh
@@ -3,4 +3,6 @@
 set -euo pipefail
 
 cargo fmt --check --all
-cargo check --workspace
+cargo check --workspace --all-targets
+cargo clippy --workspace --all-targets -- -D warnings
+cargo test --workspace --no-fail-fast
