+++ src-tauri/src/commands/development/lsp/commands.rs
@@ -20,20 +20,19 @@ use serde_json::Value;
 use std::{collections::HashMap, path::PathBuf};
 use tauri::State;
 
+type LspLaunchRequest = (
+   Option<String>,
+   Option<Vec<String>>,
+   Option<HashMap<String, String>>,
+);
+
 fn resolve_lsp_launch_request(
    app_handle: &AppHandle,
    language_id: Option<String>,
    server_path: Option<String>,
    server_args: Option<Vec<String>>,
    tools: Option<LanguageToolConfigSet>,
-) -> Result<
-   (
-      Option<String>,
-      Option<Vec<String>>,
-      Option<HashMap<String, String>>,
-   ),
-   String,
-> {
+) -> Result<LspLaunchRequest, String> {
    let Some(language_id) = language_id else {
       return Ok((server_path, server_args, None));
    };
@@ -82,6 +81,7 @@ fn locations_from_goto_response(response: Option<GotoDefinitionResponse>) -> Opt
 }
 
 #[tauri::command]
+#[allow(clippy::too_many_arguments)]
 pub async fn lsp_start(
    app_handle: AppHandle,
    lsp_manager: State<'_, LspManager>,
@@ -123,6 +123,7 @@ pub fn lsp_stop(lsp_manager: State<'_, LspManager>, workspace_path: String) -> L
 }
 
 #[tauri::command]
+#[allow(clippy::too_many_arguments)]
 pub async fn lsp_start_for_file(
    app_handle: AppHandle,
    lsp_manager: State<'_, LspManager>,
