+++ src-tauri/src/commands/ui/theme.rs
@@ -31,15 +31,14 @@ fn get_system_theme_sync() -> String {
       if let Ok(output) = Command::new("gsettings")
          .args(["get", "org.gnome.desktop.interface", "color-scheme"])
          .output()
+         && output.status.success()
       {
-         if output.status.success() {
-            let stdout = String::from_utf8_lossy(&output.stdout);
-            let theme = stdout.trim().replace('\'', "").replace('\"', "");
-            match theme.as_str() {
-               "prefer-dark" => return "dark".to_string(),
-               "prefer-light" => return "light".to_string(),
-               _ => {} // Continue to fallback
-            }
+         let stdout = String::from_utf8_lossy(&output.stdout);
+         let theme = stdout.trim().replace(['\'', '\"'], "");
+         match theme.as_str() {
+            "prefer-dark" => return "dark".to_string(),
+            "prefer-light" => return "light".to_string(),
+            _ => {} // Continue to fallback
          }
       }
 
@@ -47,12 +46,11 @@ fn get_system_theme_sync() -> String {
       if let Ok(output) = Command::new("gsettings")
          .args(["get", "org.gnome.desktop.interface", "gtk-theme"])
          .output()
+         && output.status.success()
       {
-         if output.status.success() {
-            let stdout = String::from_utf8_lossy(&output.stdout).to_lowercase();
-            if stdout.contains("dark") || stdout.contains("adwaita-dark") {
-               return "dark".to_string();
-            }
+         let stdout = String::from_utf8_lossy(&output.stdout).to_lowercase();
+         if stdout.contains("dark") || stdout.contains("adwaita-dark") {
+            return "dark".to_string();
          }
       }
 
@@ -67,12 +65,11 @@ fn get_system_theme_sync() -> String {
             "ColorScheme",
          ])
          .output()
+         && output.status.success()
       {
-         if output.status.success() {
-            let stdout = String::from_utf8_lossy(&output.stdout).to_lowercase();
-            if stdout.contains("dark") || stdout.contains("breeze dark") {
-               return "dark".to_string();
-            }
+         let stdout = String::from_utf8_lossy(&output.stdout).to_lowercase();
+         if stdout.contains("dark") || stdout.contains("breeze dark") {
+            return "dark".to_string();
          }
       }
