+++ src-tauri/src/commands/version_control/github.rs
@@ -282,6 +282,7 @@ pub async fn github_unlock_issue(
 }
 
 #[tauri::command]
+#[allow(clippy::too_many_arguments)]
 pub async fn github_create_pull_request(
    app: crate::app_runtime::AppHandle,
    repo_path: String,
